feat(MovieDetailModal): 별점 평가 기능 추가
- 사용자가 선택한 별점을 로컬 스토리지에 저장 (사용자가 별을 클릭할 때마다 해당 영화의 ID와 함께 평가 값이 업데이트 된다.)

- 영화 상세 모달에 현재 별점을 표시하는 기능을 추가 (해당 영화에 대한 이전 사용자 평가가 있을 경우, 로컬 스토리지에서 평가 값을 불러와서 표시)
